extract plots tab types
